TIP-843: Move reader to enrichment
