+++ src/Akeneo/Pim/Enrichment/Bundle/DependencyInjection/AkeneoPimEnrichmentExtension.php
@@ -83,6 +83,7 @@ class AkeneoPimEnrichmentExtension extends Extension
         $loader->load('array_converter.yml');
         $loader->load('job_parameters.yml');
         $loader->load('processors.yml');
+        $loader->load('readers.yml');
 
         if (!$container->hasParameter('pim_pdf_generator_font')) {
             $container->setParameter('pim_pdf_generator_font', null);
