+++ src/Akeneo/Pim/Enrichment/Component/Category/Connector/Reader/Database/CategoryReader.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Pim\Component\Connector\Reader\Database;
+namespace Akeneo\Pim\Enrichment\Component\Category\Connector\Reader\Database;
 
 use Akeneo\Tool\Component\Batch\Item\InitializableInterface;
 use Akeneo\Tool\Component\Batch\Item\ItemReaderInterface;
 use Akeneo\Tool\Component\Batch\Step\StepExecutionAwareInterface;
 use Akeneo\Tool\Component\Classification\Repository\CategoryRepositoryInterface;
+use Pim\Component\Connector\Reader\Database\AbstractReader;
 
 /**
  * Category reader that reads categories ordered by tree and order inside the tree
