+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Reader/File/Csv/ProductReader.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Pim\Component\Connector\Reader\File\Csv;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\File\Csv;
 
 use Pim\Component\Connector\ArrayConverter\ArrayConverterInterface;
+use Pim\Component\Connector\Reader\File\Csv\Reader;
 use Pim\Component\Connector\Reader\File\FileIteratorFactory;
 use Pim\Component\Connector\Reader\File\FileReaderInterface;
 use Pim\Component\Connector\Reader\File\MediaPathTransformer;
