+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Reader/File/Xlsx/ProductReader.php
@@ -1,31 +1,33 @@
 <?php
 
-declare(strict_types=1);
-
-namespace Pim\Component\Connector\Reader\File\Csv;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Reader\File\Xlsx;
 
 use Pim\Component\Connector\ArrayConverter\ArrayConverterInterface;
 use Pim\Component\Connector\Reader\File\FileIteratorFactory;
 use Pim\Component\Connector\Reader\File\FileReaderInterface;
 use Pim\Component\Connector\Reader\File\MediaPathTransformer;
+use Pim\Component\Connector\Reader\File\Xlsx\Reader;
 
 /**
- * Product model Csv reader
+ * Product XLSX reader
  *
- * @author    Arnaud Langlade <arnaud.langlade@akeneo.com>
- * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
+ * This specialized XLSX reader exists to replace relative media path to absolute path, in order for later process to
+ * know where to find the files.
+ *
+ * @author    Marie Bochu <marie.bochu@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductModelReader extends Reader implements FileReaderInterface
+class ProductReader extends Reader implements FileReaderInterface
 {
     /** @var MediaPathTransformer */
-    private $mediaPathTransformer;
+    protected $mediaPathTransformer;
 
     /**
      * @param FileIteratorFactory     $fileIteratorFactory
      * @param ArrayConverterInterface $converter
-     * @param MediaPathTransformer    $mediaPathTransformer
      * @param array                   $options
+     * @param MediaPathTransformer    $mediaPathTransformer
      */
     public function __construct(
         FileIteratorFactory $fileIteratorFactory,
@@ -58,21 +60,22 @@ class ProductModelReader extends Reader implements FileReaderInterface
     /**
      * @return array
      */
-    protected function getArrayConverterOptions(): array
+    protected function getArrayConverterOptions()
     {
         $jobParameters = $this->stepExecution->getJobParameters();
 
         return [
             // for the array converters
-            'mapping' => [
-                $jobParameters->get('familyVariantColumn') => 'family_variant',
+            'mapping'           => [
+                $jobParameters->get('familyColumn')     => 'family',
                 $jobParameters->get('categoriesColumn') => 'categories',
+                $jobParameters->get('groupsColumn')     => 'groups'
             ],
             'with_associations' => false,
 
             // for the delocalization
             'decimal_separator' => $jobParameters->get('decimalSeparator'),
-            'date_format' => $jobParameters->get('dateFormat'),
+            'date_format'       => $jobParameters->get('dateFormat')
         ];
     }
 }
